requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==5.2.6
django-extensions==4.1
django-filter==25.1
django-freeze==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extensions==0.8.0
drf-nested-routers==0.94.2
idna==3.10
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
xmltodict==0.14.2
